Move MockExchange separately.
